refactor: turn edit and save functions into HOCs
