+++ src/MultiFormGoals/resources/js/blocks/multi-form-goal/index.js
@@ -3,14 +3,13 @@
  */
 const { __ } = wp.i18n;
 const { registerBlockType } = wp.blocks;
-const { InnerBlocks } = wp.blockEditor;
-const { useEffect } = wp.element;
-const { select, dispatch } = wp.data;
 
 /**
  * Internal dependencies
  */
 import GiveLogo from '../../components/logo';
+import edit from './edit';
+import save from './save';
 
 /**
  * Required styles (both common and editor styles)
@@ -21,19 +20,6 @@ import '../../../css/editor.scss';
 /**
  * Register Block
  */
-const blockTemplate = [
-	[ 'core/media-text', {
-		imageFill: true,
-	}, [
-		[ 'core/heading', {
-			placeholder: __( 'Heading', 'give' ),
-		} ],
-		[ 'core/paragraph', {
-			placeholder: __( 'Summary', 'give' ),
-		} ],
-	] ],
-	[ 'give/progress-bar', {} ],
-];
 
 export default registerBlockType( 'give/multi-form-goal', {
 	title: __( 'Multi-Form Goal', 'give' ),
@@ -49,34 +35,6 @@ export default registerBlockType( 'give/multi-form-goal', {
 			'wide',
 		],
 	},
-	edit: ( { isSelected, clientId } ) => {
-		// When adding a new Multi-Form Goal block, select the inner Progress Bar block by default
-		useEffect( () => {
-			if ( isSelected ) {
-				selectProgressBar();
-			}
-		}, [] );
-
-		const selectProgressBar = () => {
-			const parentBlock = select( 'core/editor' ).getBlocksByClientId( clientId )[ 0 ];
-			const progressBarBlock = parentBlock.innerBlocks[ parentBlock.innerBlocks.length - 1 ];
-			dispatch( 'core/block-editor' ).selectBlock( progressBarBlock.clientId );
-		};
-
-		return (
-			<div className="give-multi-form-goal-block">
-				<InnerBlocks
-					template={ blockTemplate }
-					templateLock="all"
-				/>
-			</div>
-		);
-	},
-	save: () => {
-		return (
-			<div className="give-multi-form-goal-block">
-				<InnerBlocks.Content />
-			</div>
-		);
-	},
+	edit: edit,
+	save: save,
 } );
