+++ src/MultiFormGoals/resources/js/blocks/multi-form-goal/save/index.js
@@ -1,0 +1,10 @@
+const { InnerBlocks } = wp.blockEditor;
+
+const save = () => {
+	return (
+		<div className="give-multi-form-goal-block">
+			<InnerBlocks.Content />
+		</div>
+	);
+};
+export default save;
